backend/requirements.txt
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Telegram
python-telegram-bot>=20.7

# OCR & Image Processing
pytesseract>=0.3.10
opencv-python-headless>=4.9.0
Pillow>=10.2.0

# Export
openpyxl>=3.1.2

# Utils
python-dateutil>=2.8.2
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Logging & Monitoring
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Dev
black>=24.1.0
ruff>=0.1.14
